+++ include/fwd/FunctionDefinition.fwd.hpp
@@ -1,0 +1,9 @@
+#ifndef FUNCTION_DEFINITION_FWD_HPP
+#define FUNCTION_DEFINITION_FWD_HPP
+
+#include "CharType.hpp"
+
+template<CharType T>
+class FunctionDefinition;
+
+#endif
