+++ include/fwd/IExpression.fwd.hpp
@@ -1,0 +1,9 @@
+#ifndef IEXPRESSION_FWD_HPP
+#define IEXPRESSION_FWD_HPP
+
+#include "CharType.hpp"
+
+template<CharType T>
+class IExpression;
+
+#endif
