+++ include/fwd/IInstruction.fwd.hpp
@@ -1,0 +1,9 @@
+#ifndef IINSTRUCTION_FWD_HPP
+#define IINSTRUCTION_FWD_HPP
+
+#include "CharType.hpp"
+
+template<CharType T>
+class IInstruction;
+
+#endif
