+++ include/fwd/IVisitable.fwd.hpp
@@ -1,0 +1,8 @@
+#ifndef IVISITABLE_FWD_HPP
+#define IVISITABLE_FWD_HPP
+
+#include "CharType.hpp"
+template<CharType T = char>
+class IVisitable;
+
+#endif
