+++ include/fwd/IVisitor.fwd.hpp
@@ -1,0 +1,9 @@
+#ifndef IVISITOR_FWD_HPP
+#define IVISITOR_FWD_HPP
+
+#include "CharType.hpp"
+
+template<CharType T>
+class IVisitor;
+
+#endif
