+++ src/lib/ProgramTreePrinter.cpp
@@ -1,6 +1,7 @@
 #include "ProgramTreePrinter.hpp"
 
 #include <variant>
+#include "overload.hpp"
 
 template <CharType T>
 void ProgramTreePrinter<T>::print_program(const Program<T> &program){
